# api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Find name or address associated to ANS",
        "title": "ANS Indexer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.ans.notrustverify.ch",
    "basePath": "/",
    "paths": {
        "/address": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name"
                ],
                "summary": "Get names associated to an address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must fulfilled reverse and forward link",
                        "name": "strict",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Address to reverse the name",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Name"
                            }
                        }
                    }
                }
            }
        },
        "/name": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "name"
                ],
                "summary": "Get address associated to a name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name to resolve",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Must fulfilled reverse and forward link",
                        "name": "strict",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Name": {
            "type": "object",
            "properties": {
                "capitalisation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}